modularize the code
