+++ state-manager-react/src/reducers/index.js
@@ -1,0 +1,10 @@
+import { combineReducers } from "@redux-like/store";
+import counterReducer from "./counterReducer";
+import itemsReducer from "./itemsReducer";
+
+const rootReducer = combineReducers({
+  count: counterReducer,
+  items: itemsReducer,
+});
+
+export default rootReducer;
